+++ project/app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
         </activity>
 
         <activity
-            android:name=".OrganizerActivity"
+            android:name=".EventListActivity"
             android:exported="true" />
 
         <meta-data
