+++ project/app/src/main/java/com/example/fire_and_based/EventListActivity.java
@@ -1,0 +1,57 @@
+package com.example.fire_and_based;
+
+import android.os.Bundle;
+import android.view.MenuItem;
+import android.widget.ListView;
+import androidx.appcompat.widget.Toolbar;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.drawerlayout.widget.DrawerLayout;
+
+import java.util.ArrayList;
+
+public class EventListActivity extends AppCompatActivity {
+    private ArrayList<Event> dataList;
+    private ListView eventList;
+    private EventArrayAdapter eventAdapter;
+    private DrawerLayout drawerLayout;
+    private ActionBarDrawerToggle actionBarDrawerToggle;
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.events_list);
+
+        dataList = new ArrayList<Event>();
+        dataList.add(new Event("Event 1", "1"));
+        dataList.add(new Event("Event 2", "2"));
+        dataList.add(new Event("Event 3", "3"));
+        eventList = findViewById(R.id.event_list);
+        eventAdapter = new EventArrayAdapter(this, dataList);
+        eventList.setAdapter(eventAdapter);
+
+        drawerLayout = findViewById(R.id.drawer_layout);
+        actionBarDrawerToggle = new ActionBarDrawerToggle(this, drawerLayout, R.string.nav_open, R.string.nav_close);
+
+
+        drawerLayout.addDrawerListener(actionBarDrawerToggle);
+        actionBarDrawerToggle.syncState();
+
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setHomeAsUpIndicator(R.drawable.baseline_menu_24);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        if (actionBarDrawerToggle.onOptionsItemSelected(item)) {
+            return true;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+}
