+++ project/app/src/main/java/com/example/fire_and_based/MainActivity.java
@@ -6,7 +6,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Button;
 import android.view.View;
-import android.widget.Button;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -17,7 +16,7 @@ public class MainActivity extends AppCompatActivity {
 
         Button button = findViewById(R.id.button2);
         button.setOnClickListener(v -> {
-            Intent intent = new Intent(MainActivity.this, OrganizerActivity.class);
+            Intent intent = new Intent(MainActivity.this, EventListActivity.class);
             startActivity(intent);
         });
