+++ project/app/src/main/res/menu/attendee_event_menu.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/scan_qrcode_button"
+        android:title="Scan QRCode" />
+    <item
+        android:id="@+id/browse_events_button"
+        android:title="Browse Events" />
+    <item
+        android:id="@+id/edit_profile_button"
+        android:title="Edit Profile" />
+    <item
+        android:id="@+id/switch_mode_button"
+        android:title="Switch to Organizer Mode" />
+</menu>
