+++ project/app/src/main/res/values/strings.xml
@@ -94,4 +94,6 @@
     <string name="title_activity_main2">MainActivity2</string>
     <string name="tab_text_1">Tab 1</string>
     <string name="tab_text_2">Tab 2</string>
+    <string name="nav_open">Open</string>
+    <string name="nav_close">Close</string>
 </resources>
